pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "localsmartz"
version = "0.1.0"
description = "Local-first multi-agent research system powered by Ollama"
requires-python = ">=3.12"
license = {file = "LICENSE"}

dependencies = [
    "deepagents",
    "langchain-ollama",
    # Durable checkpointer — SQLite-backed threads survive backend restart.
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langsmith>=0.3.0",
    # Document processing
    "pypdf>=4.3.0",
    "pdfplumber>=0.11.0",
    "openpyxl>=3.1.4",
    "python-docx>=1.1.0",
    # Web scraping
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "httpx>=0.28.0",
    # Web search — `duckduckgo-search` was renamed to `ddgs` upstream (2025).
    # Same DDGS class, same .text(query, max_results=N) surface.
    "ddgs>=9.0",
    # OpenTelemetry — opt-in via --observe flag, but always installed so the
    # bundled DMG works without an extra pip step.
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
    "openinference-instrumentation-langchain>=0.1.0",
    # Secret storage — Keychain on macOS, secret-tool on Linux, file fallback.
    "keyring>=24.0.0",
    # Cloud provider SDKs (Phase 2). Imports are guarded in
    # runners/__init__.py so an offline wheel without these still works.
    "anthropic>=0.40",
    "openai>=1.50",
    # LangChain adapters so DeepAgents / create_deep_agent can accept
    # cloud providers via provider-swap dispatch. Anthropic pulls
    # in langchain-anthropic; OpenAI/Groq share langchain-openai with a
    # different base_url. Both are imported lazily under provider
    # branches so offline installs still resolve.
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
]

[project.scripts]
localsmartz = "localsmartz.__main__:main"
localsmartz-benchmark = "localsmartz.benchmarking:main"

[tool.hatch.build.targets.wheel]
packages = ["src/localsmartz"]

[dependency-groups]
dev = [
    "arize-phoenix>=14.5.0",
    "pytest>=9.0.3",
]
